+++ Editor/ValidationSuite/ValidationTests/ApiValidation.cs
@@ -18,6 +18,7 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests
             TestDescription = "Checks public API for style and changest that conflict with Semantic Versioning.";
             TestCategory = TestCategory.ApiValidation;
             SupportedValidations = new[] { ValidationType.CI, ValidationType.LocalDevelopmentInternal, ValidationType.Publishing };
+            CanUseValidationExceptions = true;
         }
 
         public ApiValidation(ValidationAssemblyInformation validationAssemblyInformation)